Add editButton
