+++ package.json
@@ -41,6 +41,7 @@
   },
   "dependencies": {
     "babel": "^6.23.0",
-    "date-fns": "^2.21.3"
+    "date-fns": "^2.21.3",
+    "populate.js": "^1.2.2"
   }
 }
